Split Salt managed known_hosts into sites
As we do not have a shared Salt cache implementation, each master only
has access to mine data (which we use for collecting the minion host
keys) of directly connected minions. This also means that currently the
tooling was only useful on witch1, where it was originally used, as the
commit and push on other syndic masters would lead to a diff replacing
all existing entries in the known_hosts file and ultimately a conflict.
Splitting into multiple known_hosts is not perfect, as it requires the
user to keep their "UserKnownHostsFile" setting in ssh_config(5) updated
to reflect all site-based known_hosts files, but given we do not add new
sites often, it is deemed a minor inconvenience (plus, we still offer
transparent SSHFP based host key validation as an alternative to the
file based one).

Signed-off-by: Georg Pfuetzenreuter <mail@georg-pfuetzenreuter.net>