Overwrite zypp releasever variable on Tumbleweed
On Tumbleweed, $releasever resolves to the snapshot date by default,
which in contrast to $releasever on Leap (which resolves to the dotted
Leap release) does not match the standard repository naming. So far, we
used "$releasever" in repository URLs for roles used on Leap machines,
and "openSUSE_Tumbleweed" in repository URLs for roles used on
Tumbleweed machines for this reason. Now, if a role is to be used with
both Leap and tumbleweed machines, this would require pillar templating
to set the URL path depending on the operating system. Avoid this and
simplify our setup by allowing all repository paths to be specified
using the "$releasever" variable regardless of which machines they
will be applied to by "fixing" the variable on Tumbleweed. So far,
no functionality where the default date based $releasever value would
be expected was found.

Signed-off-by: Georg Pfuetzenreuter <mail@georg-pfuetzenreuter.net>